+++ packages/app/src/cli/models/extensions/ui-specifications/checkout_ui_extension.ts
@@ -1,6 +1,6 @@
 import {createUIExtensionSpecification} from '../ui.js'
 import {BaseUIExtensionSchema} from '../schemas.js'
-import {loadLocalesConfig} from '../../../public/utilities/extensions/locales-configuration.js'
+import {loadLocalesConfig} from '../../../utilities/extensions/locales-configuration.js'
 import {schema} from '@shopify/cli-kit'
 
 const dependency = {name: '@shopify/checkout-ui-extensions-react', version: '^0.21.2'}
