+++ packages/app/src/cli/services/dev/extension/payload.ts
@@ -1,7 +1,7 @@
 import {getLocalization} from './localization.js'
 import {UIExtensionPayload} from './payload/models.js'
 import {getExtensionPointTargetSurface} from './utilities.js'
-import {getUIExtensionResourceURL} from '../../../public/utilities/extensions/configuration.js'
+import {getUIExtensionResourceURL} from '../../../utilities/extensions/configuration.js'
 import {ExtensionDevOptions} from '../extension.js'
 import {UIExtension} from '../../../models/app/extensions.js'
 import {getUIExtensionRendererVersion} from '../../../models/app/app.js'
