+++ packages/app/src/cli/utilities/extensions/fetch-product-variant.ts
