+++ packages/app/src/cli/utilities/extensions/locales-configuration.test.ts
