+++ packages/app/src/cli/utilities/extensions/locales-configuration.ts
